use a bash prompt for troubleshooting
I ran into a problem where if my test invoked zsh twice in the same
run, then second invocation would just hang. I couldn't break out of
it any way but killing the terminal window it was running in.

Running vanilla bash with --norc seems to be pretty safe in this
regard.