create refs so that pruning won't remove obsolete commits
Since git internals don't work with obsolescence graphs, obsolete
commits are seen as orphaned. This commit creates a refs namespace,
refs/cheads (for change heads), to record refs that pin them to the
repository. The plan is also to use these refs for pushing and
fetching obsolete commits in change ids in the future.

This commit introduces a shell function for calculated the change id
of any commit as a dependency for the above. See the commit for
details.