fetch obsolete intermediate commits from remote
This ones git fetch to fetch the intended remotes. Then, it looks at
each new commit and traverses the obsolescence graph until it
discovers intermediate commits that are referenced but not available
in the local repository. It builds a list of all of these missing
commits and runs fetch again to get them.

It cannot be sure that it got them all in one go since newly fetched
commits might reference others. So, it repeats the process until it
can traverse the entire graph.