use the repository specified on the fetch command line
I thought it was convenient that the fetched url could be found in
FETCH_HEAD. That meant that I didn't have to parse arguments to figure
out which argument was it. However, I found that if the URL had a user
specified (user@) then it was not found in the file.

Still trying to avoid parsing fetch options, this compares each
argument with the URL found in the file to try to figure it out
heuristically. I don't love it but it seems to work.